# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogState": {
        "enum": [
          "ready",
          "missing",
          "generating",
          "invalid"
        ],
        "title": "CatalogState",
        "type": "string"
      },
...
      "ExploreFacets": {
        "description": "Counts per facet value, each computed with every other active filter applied.",
        "properties": {
          "category": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "title": "Category",
            "type": "array"
          },
          "inheritance": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "title": "Inheritance",
            "type": "array"
          },
          "protein_family": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "title": "Protein Family",
            "type": "array"
          },
          "structure": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "title": "Structure",
            "type": "array"
          },
          "subcategory": {
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            },
            "title": "Subcategory",
            "type": "array"
          }
        },
        "required": [
          "category",
          "subcategory",
          "inheritance",
          "protein_family",
          "structure"
        ],
        "title": "ExploreFacets",
        "type": "object"
      },
      "ExploreFacetsResponse": {
        "properties": {
          "catalog_state": {
            "$ref": "#/components/schemas/CatalogState"
          },
          "facets": {
            "$ref": "#/components/schemas/ExploreFacets"
          },
          "ranges": {
            "$ref": "#/components/schemas/ExploreRanges"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "title": "Sources",
            "type": "array"
          },
          "total": {
            "description": "Genes matching the active filters",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "facets",
          "ranges",
          "total",
          "catalog_state",
          "sources"
        ],
        "title": "ExploreFacetsResponse",
        "type": "object"
      },
...
      "ExploreGenesResponse": {
        "properties": {
          "catalog_state": {
            "$ref": "#/components/schemas/CatalogState"
          },
          "facets": {
            "$ref": "#/components/schemas/ExploreFacets"
          },
          "has_more": {
            "readOnly": true,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExploreGene"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "order": {
            "$ref": "#/components/schemas/SortOrder"
          },
          "ranges": {
            "$ref": "#/components/schemas/ExploreRanges"
          },
          "sort": {
            "$ref": "#/components/schemas/ExploreSort"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "title": "Sources",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "sources",
          "sort",
          "order",
          "facets",
          "ranges",
          "catalog_state",
          "has_more"
        ],
        "title": "ExploreGenesResponse",
        "type": "object"
      },
      "ExploreRanges": {
        "properties": {
          "pathogenic_variants": {
            "$ref": "#/components/schemas/NumericRange"
          },
          "publications": {
            "$ref": "#/components/schemas/NumericRange"
          },
          "total_variants": {
            "$ref": "#/components/schemas/NumericRange"
          }
        },
        "required": [
          "pathogenic_variants",
          "total_variants",
          "publications"
        ],
        "title": "ExploreRanges",
        "type": "object"
      },
      "ExploreSort": {
        "enum": [
          "relevance",
          "symbol",
          "name",
          "pathogenic_variants",
          "total_variants",
          "publications",
          "experimental_structures",
          "protein_length"
        ],
        "title": "ExploreSort",
        "type": "string"
      },
      "FacetValue": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "table": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "label",
          "count",
          "parent",
          "table"
        ],
        "title": "FacetValue",
        "type": "object"
      },
...
      "FieldProblem": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "location",
          "message"
        ],
        "title": "FieldProblem",
        "type": "object"
      },
...
      "NumericRange": {
        "description": "Bounds over the whole catalog, for range controls. Null when no gene has a value.",
        "properties": {
          "known": {
            "description": "Genes for which the source reported a value",
            "title": "Known",
            "type": "integer"
          },
          "max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          }
        },
        "required": [
          "min",
          "max",
          "known"
        ],
        "title": "NumericRange",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": true,
        "description": "Body of every error response.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldProblem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "SourceState": {
        "enum": [
          "ok",
          "empty",
          "unavailable",
          "disabled_by_license",
          "not_configured"
        ],
        "title": "SourceState",
        "type": "string"
      },
      "SourceStatus": {
        "description": "One row per upstream source consulted for a response.",
        "properties": {
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "from_cache": {
            "default": false,
            "title": "From Cache",
            "type": "boolean"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "release": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stale": {
            "default": false,
            "title": "Stale",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/SourceState"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "source",
          "name",
          "state",
          "message",
          "release",
          "retrieved_at",
          "license",
          "url",
          "from_cache",
          "stale",
          "elapsed_ms"
        ],
        "title": "SourceStatus",
        "type": "object"
      },
...
      "StructureAvailability": {
        "enum": [
          "experimental",
          "predicted",
          "none",
          "unknown"
        ],
        "title": "StructureAvailability",
        "type": "string"
      },
...
      "VariantMetric": {
        "enum": [
          "pathogenic",
          "total"
        ],
        "title": "VariantMetric",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Open research platform for computational rare-disease work. Research and hypothesis generation\nonly: nothing served here is clinical advice.\n\nEvery error is problem JSON (`application/problem+json`). Aggregated responses carry\n`sources: SourceStatus[]`, so a page renders when one upstream source is down. Writes need the\n`X-OrphaFold-Workspace` header (an anonymous ID generated by the client).\n",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "OrphaFold API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/explore/facets": {
      "get": {
        "operationId": "list_facets_api_v1_explore_facets_get",
        "parameters": [
          {
            "description": "IUIS category ID, e.g. iuis-t1. Repeatable",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IUIS category ID, e.g. iuis-t1. Repeatable",
              "title": "Category"
            }
          },
          {
            "description": "IUIS subcategory ID. Repeatable",
            "in": "query",
            "name": "subcategory",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IUIS subcategory ID. Repeatable",
              "title": "Subcategory"
            }
          },
          {
            "description": "Inheritance code, e.g. AR, AD, XL. Repeatable",
            "in": "query",
            "name": "inheritance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inheritance code, e.g. AR, AD, XL. Repeatable",
              "title": "Inheritance"
            }
          },
          {
            "description": "Text in the gene symbol, gene or protein name, or disease name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text in the gene symbol, gene or protein name, or disease name",
              "title": "Q"
            }
          },
          {
            "description": "Protein family, exact. 'unknown' selects genes with none. Repeatable",
            "in": "query",
            "name": "protein_family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Protein family, exact. 'unknown' selects genes with none. Repeatable",
              "title": "Protein Family"
            }
          },
          {
            "description": "Structure availability: experimental, predicted, none, unknown. Repeatable",
            "in": "query",
            "name": "structure",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/StructureAvailability"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Structure availability: experimental, predicted, none, unknown. Repeatable",
              "title": "Structure"
            }
          },
          {
            "description": "Which ClinVar count the variant range applies to",
            "in": "query",
            "name": "variant_metric",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VariantMetric",
              "default": "pathogenic",
              "description": "Which ClinVar count the variant range applies to"
            }
          },
          {
            "description": "Minimum number of reported variants",
            "in": "query",
            "name": "variants_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum number of reported variants",
              "title": "Variants Min"
            }
          },
          {
            "description": "Maximum number of reported variants",
            "in": "query",
            "name": "variants_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of reported variants",
              "title": "Variants Max"
            }
          },
          {
            "description": "Minimum publication count",
            "in": "query",
            "name": "publications_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum publication count",
              "title": "Publications Min"
            }
          },
          {
            "description": "Maximum publication count",
            "in": "query",
            "name": "publications_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum publication count",
              "title": "Publications Max"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploreFacetsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          }
        },
        "summary": "Facet values, counts and numeric ranges for the explore filters",
        "tags": [
          "explore"
        ]
      }
    },
    "/api/v1/explore/genes": {
      "get": {
        "operationId": "list_genes_api_v1_explore_genes_get",
        "parameters": [
          {
            "description": "Default: relevance with q, otherwise symbol",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExploreSort"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default: relevance with q, otherwise symbol",
              "title": "Sort"
            }
          },
          {
            "description": "Genes with no value for the sort key come last",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SortOrder"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Genes with no value for the sort key come last",
              "title": "Order"
            }
          },
          {
            "description": "IUIS category ID, e.g. iuis-t1. Repeatable",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IUIS category ID, e.g. iuis-t1. Repeatable",
              "title": "Category"
            }
          },
          {
            "description": "IUIS subcategory ID. Repeatable",
            "in": "query",
            "name": "subcategory",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IUIS subcategory ID. Repeatable",
              "title": "Subcategory"
            }
          },
          {
            "description": "Inheritance code, e.g. AR, AD, XL. Repeatable",
            "in": "query",
            "name": "inheritance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inheritance code, e.g. AR, AD, XL. Repeatable",
              "title": "Inheritance"
            }
          },
          {
            "description": "Text in the gene symbol, gene or protein name, or disease name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text in the gene symbol, gene or protein name, or disease name",
              "title": "Q"
            }
          },
          {
            "description": "Protein family, exact. 'unknown' selects genes with none. Repeatable",
            "in": "query",
            "name": "protein_family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Protein family, exact. 'unknown' selects genes with none. Repeatable",
              "title": "Protein Family"
            }
          },
          {
            "description": "Structure availability: experimental, predicted, none, unknown. Repeatable",
            "in": "query",
            "name": "structure",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/StructureAvailability"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Structure availability: experimental, predicted, none, unknown. Repeatable",
              "title": "Structure"
            }
          },
          {
            "description": "Which ClinVar count the variant range applies to",
            "in": "query",
            "name": "variant_metric",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VariantMetric",
              "default": "pathogenic",
              "description": "Which ClinVar count the variant range applies to"
            }
          },
          {
            "description": "Minimum number of reported variants",
            "in": "query",
            "name": "variants_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum number of reported variants",
              "title": "Variants Min"
            }
          },
          {
            "description": "Maximum number of reported variants",
            "in": "query",
            "name": "variants_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of reported variants",
              "title": "Variants Max"
            }
          },
          {
            "description": "Minimum publication count",
            "in": "query",
            "name": "publications_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum publication count",
              "title": "Publications Min"
            }
          },
          {
            "description": "Maximum publication count",
            "in": "query",
            "name": "publications_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum publication count",
              "title": "Publications Max"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Items to skip",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Items to skip",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploreGenesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          }
        },
        "summary": "Filter, sort and page the IEI gene set, with facet counts",
        "tags": [
          "explore"
        ]
      }
    },
...
  }
}